requirements.txt
opencv-python
faiss-gpu
numpy
tqdm
natsort
pathlib
glob
opencv-contrib-python-headless
pickle5
matplotlib
torchvision
torchinfo
torchsummary
timm
ptflops